workers/analytics-worker/requirements.txt
nats-py
asyncpg
